server-client simple websocket test
