+++ my-node-app/public/index.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>WebSocket Test</title>
+</head>
+<body>
+    <h1>WebSocket Test Page</h1>
+    <button onclick="connectWebSocket()">Connect WebSocket</button>
+    <script src="js/websocket-test.js"></script>
+</body>
+</html>
