--- my-node-app/public/websocket-test.html
@@ -1,36 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>WebSocket Test</title>
-</head>
-<body>
-    <h1>WebSocket Test</h1>
-    <div id="output"></div>
-
-    <script>
-        const socket = new WebSocket('ws://localhost:8080');
-
-        socket.onopen = () => {
-            console.log('WebSocket connection established');
-            document.getElementById('output').innerHTML = 'WebSocket connection established';
-        };
-
-        socket.onmessage = (event) => {
-            console.log('Message from server:', event.data);
-            document.getElementById('output').innerHTML = `Message from server: ${event.data}`;
-        };
-
-        socket.onclose = () => {
-            console.log('WebSocket connection closed');
-            document.getElementById('output').innerHTML = 'WebSocket connection closed';
-        };
-
-        socket.onerror = (error) => {
-            console.error('WebSocket error:', error);
-            document.getElementById('output').innerHTML = 'WebSocket error occurred';
-        };
-    </script>
-</body>
-</html>
