--- my-node-app/utils/UPDATE current_rssi_measurements SET sen.sql
@@ -1,1 +1,0 @@
-UPDATE current_rssi_measurements SET send_flag = 0;
